+++ diffs/AaveV3Base_MayFundingPartB_20250524_before_AaveV3Base_MayFundingPartB_20250524_after.md
@@ -1,0 +1,52 @@
+## Raw diff
+
+```json
+{
+  "raw": {
+    "0x0a1d576f3efef75b330424287a95a366e8281d54": {
+      "label": "AaveV3Base.ASSETS.USDbC.A_TOKEN",
+      "balanceDiff": null,
+      "stateDiff": {
+        "0x74b7bcaf0f88f03d5090dfec8ef1441b6159a579b34e65030e319d50d2a00cf2": {
+          "previousValue": "0x0000000000000000000000000000000000000000000000000000000000000000",
+          "newValue": "0x0000000000000000000000000000000000000000000000000000000858a717ca"
+        }
+      }
+    },
+    "0x2dc219e716793fb4b21548c0f009ba3af753ab01": {
+      "label": "GovernanceV3Base.PAYLOADS_CONTROLLER",
+      "balanceDiff": null,
+      "stateDiff": {
+        "0x8195e934be8e1e1e67e36670679242f3eb3fe013d19203686902c1dc42dff3e4": {
+          "previousValue": "0x006831a82c000000000002000000000000000000000000000000000000000000",
+          "newValue": "0x006831a82c000000000003000000000000000000000000000000000000000000"
+        },
+        "0x8195e934be8e1e1e67e36670679242f3eb3fe013d19203686902c1dc42dff3e5": {
+          "previousValue": "0x000000000000000000093a80000000000000685fccad00000000000000000000",
+          "newValue": "0x000000000000000000093a80000000000000685fccad0000000000006831a82d"
+        }
+      }
+    },
+    "0x4e65fe4dba92790696d040ac24aa414708f5c0ab": {
+      "label": "AaveV3Base.ASSETS.USDC.A_TOKEN",
+      "balanceDiff": null,
+      "stateDiff": {
+        "0x74b7bcaf0f88f03d5090dfec8ef1441b6159a579b34e65030e319d50d2a00cf2": {
+          "previousValue": "0x0000000000000000000000000000000000000000000000000000000000000000",
+          "newValue": "0x000000000000000000000000000000000000000000000000000000870688c812"
+        }
+      }
+    },
+    "0xbdb9300b7cde636d9cd4aff00f6f009ffbbc8ee6": {
+      "label": "AaveV3Base.ASSETS.cbBTC.A_TOKEN",
+      "balanceDiff": null,
+      "stateDiff": {
+        "0x74b7bcaf0f88f03d5090dfec8ef1441b6159a579b34e65030e319d50d2a00cf2": {
+          "previousValue": "0x0000000000000000000000000000000000000000000000000000000000000000",
+          "newValue": "0x00000000000000000000000000000000000000000000000000000000089594ab"
+        }
+      }
+    }
+  }
+}
+```
