--- src/20250522_AaveV3Ethereum_MayFundingPartB/MayFundingPartB.md
@@ -1,22 +1,0 @@
----
-title: "May Funding Part B"
-author: "@TokenLogic"
-discussions: "https://governance.aave.com/t/arfc-may-2025-funding-update/21906"
----
-
-## Simple Summary
-
-## Motivation
-
-## Specification
-
-## References
-
-- Implementation: [AaveV3Ethereum](https://github.com/bgd-labs/aave-proposals-v3/blob/main/src/20250522_AaveV3Ethereum_MayFundingPartB/AaveV3Ethereum_MayFundingPartB_20250522.sol)
-- Tests: [AaveV3Ethereum](https://github.com/bgd-labs/aave-proposals-v3/blob/main/src/20250522_AaveV3Ethereum_MayFundingPartB/AaveV3Ethereum_MayFundingPartB_20250522.t.sol)
-- Snapshot: Direct-to-AIP
-- [Discussion](https://governance.aave.com/t/arfc-may-2025-funding-update/21906)
-
-## Copyright
-
-Copyright and related rights waived via [CC0](https://creativecommons.org/publicdomain/zero/1.0/).
